- contained NPEs case - removed redundant setters - moved "tests" from Main to Test
