+++ src/com/kopytDev/HealthyBurger.java
@@ -1,6 +1,6 @@
-package com.company;
+package com.kopytDev;
 
-public class HealthyBurger extends Burgers {
+public class HealthyBurger extends Burger {
 
     public HealthyBurger(double basePrice) {
         super("Brown", basePrice, "Healthy Burger", 6);
